add comment and split core file into types, constanst, helpers
