+++ src/components/addBlockModal/addBlockModalConstants.ts
@@ -15,32 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IBlock, RegistryId } from "@/core";
-
 export const TAG_ALL = "All Categories";
 export const BLOCK_RESULT_COLUMN_COUNT = 2;
 export const POPULAR_BRICK_TAG_ID = "35367896-b38f-447e-9444-ecfecb258468";
-
-export type BlockResult = IBlock & {
-  isPopular?: boolean;
-};
-
-export type BlockOption = {
-  blockResult: BlockResult;
-  value: RegistryId;
-  label: string;
-};
-
-export type BlockGridData = {
-  blockOptions: BlockOption[];
-  onSetDetailBlock: (block: IBlock) => void;
-  onSelectBlock: (block: IBlock) => void;
-};
-
-export function getFlatArrayIndex(
-  rowIndex: number,
-  columnIndex: number
-): number {
-  // Layout items in the grid left to right, top to bottom
-  return rowIndex * BLOCK_RESULT_COLUMN_COUNT + columnIndex;
-}
