+++ src/components/addBlockModal/groupListingsByTag.test.ts
@@ -19,7 +19,7 @@ import {
   marketplaceListingFactory,
   marketplaceTagFactory,
 } from "@/testUtils/factories";
-import { POPULAR_BRICK_TAG_ID } from "@/components/addBlockModal/addBlockModalCore";
+import { POPULAR_BRICK_TAG_ID } from "@/components/addBlockModal/addBlockModalConstants";
 import groupListingsByTag from "@/components/addBlockModal/groupListingsByTag";
 
 describe("groupListingsByTag", () => {
